app/requirements.txt
# Python dependencies for the Streamlit Databricks App
streamlit>=1.30.0,<2.0.0
plotly>=5.18.0,<6.0.0
databricks-sdk>=0.20.0,<1.0.0
pandas>=2.0.0,<3.0.0
